Fixed strange bug with using uiop:define-package and ros install.
It gave this error message:

    #<THREAD "main thread" RUNNING {1004A70143}>:
      Error finding package for symbol "40ANTS-DOC":
     The name "ROS.SCRIPT.INSTALL.3671936092" does not designate any package.

if I used

   qlot exec ros install 40ants-doc

and sebsequent:

   qlot exec ros run -L sbcl-bin --eval '(ql:quickload :40ants-doc)' --quit

The problem is in the symbolic name of the package name in the
define-package form. Seems, this symbol intered into the package which
is active at the moment when package is loaded. In case of
using ros install, it is the package of defined in the ros script:
`ROS.SCRIPT.INSTALL.3671936092`. To prevent this, a keyword or a string
should be used in the define-package form. Probably, this bug should be
fixed in the ASDF.